refatorando codigo para cc - passando funcao do filtro para dentro da pasta service
