+++ src/context/Provider.js
@@ -5,7 +5,7 @@ import Context from './Context';
 export default function Provider({ children }) {
   const [foodData, setFoodData] = useState([]);
   const [showSearchBar, setSearchBar] = useState(false);
-  
+
   function changeSearchBar(show) {
     if (show) {
       setSearchBar(false);
